third_party/WebKit/LayoutTests/bluetooth/characteristic/readValue/gen-gatt-op-garbage-collection-ran-during-error.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
let test_desc = 'Garbage Collection ran during a readValue call that '
   'fails. Should not crash.'
let characteristic, fake_characteristic, func_promise;
promise_test(() => getMeasurementIntervalCharacteristic()
  .then(_ => ({characteristic, fake_characteristic} = _))
  // 1. Set up error responses.
  .then(() => fake_characteristic.setNextReadResponse(1))
  .then(() => fake_characteristic.setNextWriteResponse(1))
  .then(() => fake_characteristic.setNextSubscribeToNotificationsResponse(1))
  .then(() => {
    // 2. Begin making a call that will fail.
    func_promise = assert_promise_rejects_with_message(
        characteristic.readValue(),
        new DOMException(
            'GATT Server is disconnected. Cannot perform GATT operations. ' +
            '(Re)connect first with `device.gatt.connect`.',
            'NetworkError'));
    // 3. Disconnect called to clear attributeInstanceMap and allow the object
    // to get garbage collected.
    characteristic.service.device.gatt.disconnect();
    characteristic = undefined;
    fake_characteristic = undefined;
  })
  // 4. Run garbage collection.
  .then(runGarbageCollection)
  .then(() => func_promise), test_desc);

</script>
